% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CARIRT.R
\name{CARIRT}
\alias{CARIRT}
\title{A fast C++ implementation of the conditional autoregressive scaling model}
\usage{
CARIRT(ConnectionMatrix, VoteMatrix, IRT.polarity = 1, burn.iter = 100,
  mcmc.iter = 100, thin.iter = 1, store.alpha = FALSE,
  store.beta = FALSE, store.lambda = FALSE, store.average.pref = FALSE,
  store.latent.pref = FALSE, store.expected.pref = FALSE)
}
\arguments{
\item{ConnectionMatrix}{}

\item{VoteMatrix}{}

\item{IRT.polarity}{}

\item{burn.iter}{}

\item{mcmc.iter}{}

\item{thin.iter}{}

\item{store.alpha}{}

\item{store.beta}{}

\item{store.lambda}{}

\item{store.average.pref}{}

\item{store.latent.pref}{}

\item{store.expected.pref}{}
}
\description{
\code{CARIRT}
}

