\name{eliot-package}
\alias{eliot-package}
\docType{package}
\title{
Miscellaneous R Functions
}
\description{
A collection of R functions created by Benjamin E. Lauderdale.  Includes some replication data.}
\details{
\tabular{ll}{
Package: \tab eliot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-28\cr
License: \tab GPL (>= 2)\cr
}

Time for you and time for me,\cr
And time yet for a hundred indecisions,\cr
And for a hundred visions and revisions,\cr
Before the taking of a toast and tea.\cr
\cr
\emph{The Love Song of J. Alfred Prufrock}, T.S. Eliot

}
\author{
Benjamin E. Lauderdale \url{b.e.lauderdale@lse.ac.uk}
}
\keyword{ package }

