% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{emptyplot}
\alias{emptyplot}
\title{Create empty plot.}
\usage{
emptyplot(xlim = c(-1, 1), ylim = c(-1, 1), xlab = "", ylab = "",
  main = "", sub = "", ...)
}
\arguments{
\item{p}{}
}
\description{
\code{emptyplot}.
}

