% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{fadecol}
\alias{fadecol}
\title{Function to fade colours by specified alpha}
\usage{
fadecol(colours, alpha = 1)
}
\arguments{
\item{colours}{a vector providing any of the three kinds of R color specifications: hex, character string, or integer.}

\item{alpha}{the desired alpha level, must be in the interval 0 to 1.}
}
\description{
\code{fadecol} takes existing colours and applies user-specified alpha levels to them.
}

