% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LDAIRT.R
\name{LDAIRT}
\alias{LDAIRT}
\title{A fast C++ implementation of the LDA + IRT scaling model}
\usage{
LDAIRT(WordSparseMatrix, VoteMatrix, LDA.topics = 10,
  LDA.alpha = 1/LDA.topics, LDA.beta = 1, IRT.discrim.prec = 0.25,
  IRT.polarity = 1, burn.iter = 100, mcmc.iter = 100,
  save.word.chain = FALSE, save.topic.chain = FALSE,
  save.ideal.chain = TRUE, save.discrim.chain = FALSE)
}
\arguments{
\item{WordSparseMatrix}{}

\item{VoteMatrix}{}

\item{LDA.topics}{}

\item{LDA.alpha}{}

\item{LDA.beta}{}

\item{IRT.discrim.prec}{}

\item{IRT.polarity}{}

\item{burn.iter}{}

\item{mcmc.iter}{}

\item{save.word.chain}{}

\item{save.topic.chain}{}

\item{save.ideal.chain}{}

\item{save.discrim.chain}{}
}
\description{
\code{LDAIRT}
}

