% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/round.R
\name{round1}
\alias{round1}
\title{Convenience function to round values for display.}
\usage{
round1(x)
}
\arguments{
\item{x}{}
}
\description{
\code{round1}, \code{round2}, and \code{round2} round to 1, 2, and 3 points after the decimal, using the \code{formatC} function to retain trailing zeroes.
}

