% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{shadeBands}
\alias{shadeBands}
\title{Add shaded confidence/prediction bands to a plot.}
\usage{
shadeBands(x, lo, hi, ...)
}
\arguments{
\item{x}{values of x at which interval has been calculated}

\item{lo}{lower bound of interval, at each value of x}

\item{hi}{upper bound of interval, at each value of x}

\item{...}{additional arguments for the polygon function.}
}
\description{
\code{shadeBands}.
}

