% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{standardize}
\alias{standardize}
\title{Function to standardize a vector}
\usage{
standardize(x, na.rm = FALSE)
}
\arguments{
\item{x}{}
}
\description{
\code{standardize} subtracts the mean from and divides by the standard deviation of a vector.
}

