% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.LDAIRT.R
\name{summary.LDAIRT}
\alias{summary.LDAIRT}
\title{A summary method for the LDAIRT model output}
\usage{
\method{summary}{LDAIRT}(LDAIRT.out, print.topwords = TRUE,
  print.topvotes = TRUE, print.ordering = FALSE)
}
\arguments{
\item{LDAIRT.out}{}

\item{print.topwords}{}

\item{print.topvotes}{}

\item{print.ordering}{}
}
\description{
\code{summary.LDAIRT}
}

