% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordminnow.R
\name{wordminnow}
\alias{wordminnow}
\title{A fast C++ implementation of the Wordfish text model}
\usage{
wordminnow(wfm, dir = c(1, 10), control = list(priors = c(Inf, Inf, 3, 1),
  tol = c(1e-06, 1e-08)))
}
\arguments{
\item{wfm}{The word/feature frequency matrix on which the model will be fit.}

\item{dir}{}

\item{priors}{}

\item{tol}{}
}
\description{
\code{wordminnow} implements Slapin and Proksch's Wordfish model in C++ for faster implementation.  Based on Will Lowe's EM algorithm R code in the package \code{austin}.
}

